add field "extensions" in object "resolve" for file "webpack.config.js"
